Refactor `PropertiesNode` dispatches
